# k8s-deploy/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a KubeFATE.",
        "title": "KubeFATE service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/FederatedAI/KubeFATE"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v1"
    },
    "host": "kubefate.net",
    "basePath": "/v1",
    "paths": {
...
        "/log/{clusterID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "When the container of requestArgs is not set, the logs of all components will be obtained",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Log"
                ],
                "summary": "Get cluster log by clusterID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cluster ID",
                        "name": "clusterID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "container name",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "previous",
                        "name": "previous",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "since",
                        "name": "since",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "since-time",
                        "name": "since-time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "timestamps",
                        "name": "timestamps",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "tail",
                        "name": "tail",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "limit-bytes",
                        "name": "limit-bytes",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    }
                }
            }
        },
        "/log/{clusterID}/ws": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Log"
                ],
                "summary": "Get Cluster Log flow (http1.1)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cluster ID",
                        "name": "clusterID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "container name",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "previous",
                        "name": "previous",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "since",
                        "name": "since",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "since-time",
                        "name": "since-time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "timestamps",
                        "name": "timestamps",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "tail",
                        "name": "tail",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "limit-bytes",
                        "name": "limit-bytes",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.JSONERRORResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.JSONResult": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}